chore: simplify namespaces and remove redundant usings across the codebase
