+++ examples/Operator/Program.cs
@@ -5,6 +5,7 @@
 #if DEBUG
 using KubeOps.Abstractions.Crds;
 #endif
+
 using KubeOps.Operator;
 
 using Microsoft.Extensions.Hosting;
