+++ src/KubeOps.Aspire/KubeOpsServiceDefaultsExtensions.cs
@@ -4,7 +4,6 @@
 
 using KubeOps.Abstractions.Builder;
 
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Hosting;
