+++ src/KubeOps.Cli/Commands/Management/Uninstall.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.CommandLine;
-using System.CommandLine.Invocation;
 
 using k8s;
 using k8s.Autorest;
