+++ src/KubeOps.Generator/Discovery/AttributedEntity.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-namespace KubeOps.Generator.SyntaxReceiver;
+namespace KubeOps.Generator.Discovery;
 
 internal record struct AttributedEntity(
     ClassDeclarationMetaData ClassDeclaration,
