+++ src/KubeOps.Operator/Events/EventNameEncoder.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Security.Cryptography;
-
 using System.Text;
 
 namespace KubeOps.Operator.Events;
