+++ test/KubeOps.Operator.Test/Events/EventPublisher.Integration.Test.cs
@@ -2,9 +2,6 @@
 // The .NET Foundation licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-using System.Security.Cryptography;
-using System.Text;
-
 using FluentAssertions;
 
 using k8s.Models;
